Feat: "수정 회원정보 수정 API, 닉네임 중복 확인 API, 회원 프로필 사진 업로드 API"
- 로그인한 사용자만 접근 가능하도록 수정
- 그로인한 URL주소 변경
- 기존 URL로 memberId를 받아왔는데, 이젠 로그인한 사용자의 토큰을 통해서 해당 로그인 사용자의 회원정보만을 수정 가능
- 해당 API 사용시 회원탈퇴유무 확인

+++ src/main/java/me/coldrain/ninetyminute/service/MemberService.java
@@ -10,6 +10,7 @@ import me.coldrain.ninetyminute.entity.MemberRoleEnum;
 import me.coldrain.ninetyminute.exception.ErrorCode;
 import me.coldrain.ninetyminute.repository.AbilityRepository;
 import me.coldrain.ninetyminute.repository.MemberRepository;
+import me.coldrain.ninetyminute.security.UserDetailsImpl;
 import me.coldrain.ninetyminute.security.jwt.JwtTokenProvider;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -38,31 +39,42 @@ public class MemberService {
     //회원가입
     @Transactional
     public ResponseEntity<?> memberSignup(MemberRegisterRequest memberRegisterRequest) {
-        Optional<Member> found = memberRepository.findByUsername(memberRegisterRequest.getEmail());
-        if (found.isPresent()) {
-            return new ResponseEntity<>("중복된 이메일입니다.", HttpStatus.BAD_REQUEST);
-        }
+        Member memberCheck = memberRepository.findByUsername(memberRegisterRequest.getEmail()).orElse(null);
 
-        if (!memberRegisterRequest.getPassword().equals(memberRegisterRequest.getConfirmpassword())) {
-            return new ResponseEntity<>("재확인 비밀번호가 다릅니다.", HttpStatus.BAD_REQUEST);
+        if (memberCheck != null) {
+            if (memberCheck.isSecessionState()) {
+                return new ResponseEntity<>("탈퇴 처리된 아이디는 30일 후에 사용 가능합니다.", HttpStatus.BAD_REQUEST);
+            } else {
+                return new ResponseEntity<>("중복된 이메일입니다.", HttpStatus.BAD_REQUEST);
+            }
         } else {
-            MemberRoleEnum role = MemberRoleEnum.USER;
+            if (!memberRegisterRequest.getPassword().equals(memberRegisterRequest.getConfirmpassword())) {
+                return new ResponseEntity<>("재확인 비밀번호가 다릅니다.", HttpStatus.BAD_REQUEST);
+            } else {
+                MemberRoleEnum role = MemberRoleEnum.USER;
 
-            Member member = new Member(memberRegisterRequest, role);
-            member.encryptPassword(passwordEncoder);
-            memberRepository.save(member);
+                Member newMember = new Member(memberRegisterRequest, role);
+                newMember.encryptPassword(passwordEncoder);
+                memberRepository.save(newMember);
 
-            return new ResponseEntity<>(HttpStatus.CREATED);
+                return new ResponseEntity<>(HttpStatus.CREATED);
+            }
         }
     }
 
     //아이디 중복 확인
     public ResponseEntity<?> duplicateCheckEmail(MemberEmailDuplicateRequest memberEmailDuplicateRequest) {
         MemberDuplicateResponse memberDuplicateResponse = new MemberDuplicateResponse();
-        Optional<Member> found = memberRepository.findByUsername(memberEmailDuplicateRequest.getEmail());
-        if (found.isPresent()) {
-            memberDuplicateResponse.setExist(true);
-            memberDuplicateResponse.setMessage("중복된 이메일입니다.");
+        Member memberCheck = memberRepository.findByUsername(memberEmailDuplicateRequest.getEmail()).orElse(null);
+
+        if (memberCheck != null) {
+            if (memberCheck.isSecessionState()) {
+                memberDuplicateResponse.setExist(true);
+                memberDuplicateResponse.setMessage("탈퇴 처리된 아이디는 30일 후에 사용 가능합니다.");
+            } else {
+                memberDuplicateResponse.setExist(true);
+                memberDuplicateResponse.setMessage("중복된 이메일입니다.");
+            }
         } else {
             memberDuplicateResponse.setExist(false);
             memberDuplicateResponse.setMessage("사용 가능한 이메일입니다.");
@@ -86,10 +98,14 @@ public class MemberService {
 
     //회원정보 수정
     @Transactional
-    public ResponseEntity<?> memberEdit(Long memberId,
+    public ResponseEntity<?> memberEdit(UserDetailsImpl userDetails,
                                         MemberEditRequest memberEditRequest) {
         try {
-            Member member = memberRepository.findById(memberId).orElseThrow();
+            Member member = memberRepository.findById(userDetails.getUser().getId()).orElseThrow();
+            if (member.isSecessionState()) {
+                return new ResponseEntity<>("탈퇴 처리된 회원입니다.", HttpStatus.BAD_REQUEST);
+            }
+
             Optional<Member> found = memberRepository.findByNickname(memberEditRequest.getNickname());
             if (found.isPresent()) {
                 return new ResponseEntity<>("중복된 닉네임입니다.", HttpStatus.BAD_REQUEST);
@@ -110,32 +126,35 @@ public class MemberService {
 
     //회원 프로필 사진 업로드
     @Transactional
-    public ResponseEntity<?> memberProFileImageEdit(Long memberId,
+    public ResponseEntity<?> memberProFileImageEdit(UserDetailsImpl userDetails,
                                                     MultipartFile proFileImage) {
         try {
-            Member member = memberRepository.findById(memberId).orElseThrow();
+            Member member = memberRepository.findById(userDetails.getUser().getId()).orElseThrow();
+            if (member.isSecessionState()) {
+                return new ResponseEntity<>("탈퇴 처리된 회원입니다.", HttpStatus.BAD_REQUEST);
+            }
 
             if (member.getProfileName() == null && (proFileImage == null || proFileImage.isEmpty())) {
                 Map<String, String> profileImg = new HashMap<>();
                 profileImg.put("url", "/images/basicprofileimage.png");
                 profileImg.put("transImgFileName", "basicImage");
-                member.memberproFileImageUpdate(profileImg);
+                member.memberProFileImageUpdate(profileImg);
 
             } else if (member.getProfileName() == null) {
                 Map<String, String> profileImg = awsS3Service.uploadFile(proFileImage);
-                member.memberproFileImageUpdate(profileImg);
+                member.memberProFileImageUpdate(profileImg);
             } else {
                 if (member.getProfileName().equals("basicImage") && (proFileImage == null || proFileImage.isEmpty())) {
                     return new ResponseEntity<>(HttpStatus.OK);
 
                 } else if (member.getProfileName().equals("basicImage")) {
                     Map<String, String> profileImg = awsS3Service.uploadFile(proFileImage);
-                    member.memberproFileImageUpdate(profileImg);
+                    member.memberProFileImageUpdate(profileImg);
 
                 } else if (!(proFileImage == null || proFileImage.isEmpty())) {
                     awsS3Service.deleteFile(member.getProfileName());
                     Map<String, String> profileImg = awsS3Service.uploadFile(proFileImage);
-                    member.memberproFileImageUpdate(profileImg);
+                    member.memberProFileImageUpdate(profileImg);
                 }
             }
             return new ResponseEntity<>(HttpStatus.OK);
@@ -152,7 +171,11 @@ public class MemberService {
         } catch (Exception e) {
             return new ResponseEntity<>(ErrorCode.USERNAME_OR_PASSWORD_NOTFOUND.getMessage(), ErrorCode.USERNAME_OR_PASSWORD_NOTFOUND.getStatus());
         }
-        Member member = memberRepository.findByUsername(memberLoginRequest.getEmail()).orElse(null);
+        Member member = memberRepository.findByUsername(memberLoginRequest.getEmail()).orElseThrow();
+        if (member.isSecessionState()) {
+            return new ResponseEntity<>("탈퇴 처리된 회원입니다.", HttpStatus.BAD_REQUEST);
+        }
+
         return new ResponseEntity<>(jwtTokenCreate(member), HttpStatus.CREATED);
     }
